Add workaround for AUFS backend issue with postgresql
Reorder layer commands in order to install ssl-cert in the first layer
in order to avoid permission denied issue when starting postgresql on
/etc/ssl/private/ssl-cert-snakeoil.key.

This is due to a limitation of the AUFS docker backend.
Ref to https://github.com/chbrandt/docker-dachs/issues/1 for
an instance of the bug and to https://github.com/docker/docker/issues/783
for more details on the AUFS issue.

Change-Id: I0afe01e880f4ace4a38d3751a8c621c97d97d658